requirements.txt
google.generativeai
pandas
pyodbc
flask